requirements.txt
python-dotenv
apscheduler
sqlalchemy
tweepy
beautifulsoup4